Move grpc.go from utils to ads package
